Updates to test and split around logic, also stop creating a new server for each rpc connection
